+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+
 ## [ Unreleased ] - ReleaseDate
 
 ### Added
@@ -10,12 +11,25 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - `mock!` and `#[automock]` now support `unsafe` traits.
   ([#313](https://github.com/asomers/mockall/pull/313))
 
+## [ 0.10.2 ] - 2021-07-12
+
+### Fixed
+
+- Fix mocking specializing methods of non-generic structs, a regression in
+  v0.10.0.
+  ([#309](https://github.com/asomers/mockall/pull/309))
+
+- Fix mocking generic methods of generic structs returning nonstatic, a
+  regression in v0.10.0.
+  ([#312](https://github.com/asomers/mockall/pull/312))
+
+
 ## [ 0.10.1 ] - 2021-07-01
 
-## Fixed
+### Fixed
 
 - Fix mocking trait methods whose return values have lifetime parameters, a
-  regression since v0.10.0.
+  regression in v0.10.0.
   ([#304](https://github.com/asomers/mockall/pull/304))
 
 ## [ 0.10.0 ] - 2021-06-27